Add files via upload
Finds all the cycles that a single vertex is a member of. Accepts command line arguments or asks for input if not given. First argument is the vertex count that you are running on, while the second is the name of a file that contains a space separated adjacency matrix.